requirements.txt
google-adk==1.14.0
google-cloud-logging
google-cloud-storage
google-genai>=1.0.0
psycopg2-binary>=2.9.9
httpx>=0.27.0
bcrypt>=4.0.0
python-dotenv>=1.0.0
uvicorn>=0.30.0
aiosqlite>=0.20.0
sqlalchemy>=2.0.0
python-multipart>=0.0.9
apscheduler>=3.10.0